significantly improved plotting speed
In the PolyPoints class (base of all other Polyxxx classes) the drawing was done by looping through the coordinates and constructing the path according to the drawstyle.
The loop has been replaced by equivalent numpy methods and the resulting path is drawn calling DrawLines. Also, the path is converted to int32 prior to calling DrawLines which further increases speed.

In order to avoid problems with implicit type conversions I changed an inplace addition by a normal one. x/yAxis and dist are coordinates in user space, so in general they are of type float.